pyproject.toml
[project]
name = "sunday-league"
version = "0.1.0"
description = ""
authors = [
    {name = "Pestov Nikita",email = "pestovgrizzly@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.1,<23.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "sunday_league", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"